+++ IGraphics/IGraphics.h
@@ -1650,7 +1650,7 @@ protected:
    * @param drawScale \todo
    * @param cacheable Used to make sure the underlying bitmap can be shared between plug-in instances
    * @return APIBitmap* The new API Bitmap */
-  virtual APIBitmap* CreateAPIBitmap(int width, int height, int scale, double drawScale, bool cacheable = false) = 0;
+  virtual APIBitmap* CreateAPIBitmap(int width, int height, float scale, double drawScale, bool cacheable = false) = 0;
 
   /** Drawing API method to load a font from a PlatformFontPtr, called internally
    * @param fontID A CString that will be used to reference the font
